fix(cloudtrail_multi_region_enabled_logging_management_events): Handle event selectors and add tests
